Modificación en la db, seguridad en sesiones

+++ ini/Tutores/problemas_api/del.php
@@ -1,15 +1,24 @@
 <?php
-    require '../../princ/conexion.php';
 
-    $id = base64_decode($_GET['w']);
+require '../../princ/Seguridad.php';
+$seguridad = new Seguridad();
 
-    $consultaDel = mysqli_query($conn,"DELETE FROM problemas WHERE idProblema = '$id'");
+if ($seguridad->getUsuario() == null) {
+    header('Location: ../../../index.html');
+    exit;
+}
 
 
-    if($consultaDel === TRUE){
-        echo "<script>alert('Problema eliminado satisfactoriamente!'); history.back(); </script>";
-    }else{
-        echo "<script>alert('Ha ocurrido un problema al eliminar!'); history.back(); </script>";
-    }
 
-?>
+require '../../princ/conexion.php';
+
+$id = base64_decode($_GET['w']);
+
+$consultaDel = mysqli_query($conn, "DELETE FROM problemas WHERE idProblema = '$id'");
+
+
+if ($consultaDel === TRUE) {
+    echo "<script>alert('Problema eliminado satisfactoriamente!'); history.back(); </script>";
+} else {
+    echo "<script>alert('Ha ocurrido un problema al eliminar!'); history.back(); </script>";
+}
